Add new docker build stage

+++ anubis/gen-protos.sh
@@ -8,14 +8,17 @@ cmd="$@"
 
 # Build all models
 # ../run.sh build //models
-bazel build //models
+if [ ! -z "$SKIPPROTOCOMPILATION" ]
+then
+    bazel build //models
 
-# Now copy all generated typescript files
-# find ../bazel-bin/models/ -type f -path '**/models/**/*.js'
-# find ../bazel-bin/models/ -type f -path '**/models/**/*.ts'
-rm ./src/generated/protobuf/*
-find ../bazel-bin/models/ -type f -path '**/models/**/*.ts' -exec /bin/cp -rf '{}' ./src/generated/protobuf ';'
-find ../bazel-bin/models/ -type f -path '**/models/**/*.js' -exec /bin/cp -rf '{}' ./src/generated/protobuf ';'
-chmod 777 -R ./src/generated/protobuf
+    # Now copy all generated typescript files
+    # find ../bazel-bin/models/ -type f -path '**/models/**/*.js'
+    # find ../bazel-bin/models/ -type f -path '**/models/**/*.ts'
+    rm ./src/generated/protobuf/*
+    find ../bazel-bin/models/ -type f -path '**/models/**/*.ts' -exec /bin/cp -rf '{}' ./src/generated/protobuf ';'
+    find ../bazel-bin/models/ -type f -path '**/models/**/*.js' -exec /bin/cp -rf '{}' ./src/generated/protobuf ';'
+    chmod 777 -R ./src/generated/protobuf
+fi
 
 exec $cmd
